Add classes for each object type
This adds type specific sub-classes of `ObjectDefinition` for each
object type.

This lets us add the `identifier` property that returns a string value,
that together with the object type is a unique identifier for that
object. Object types that cannot be uniquely identified based on their
data will return `None`.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>